+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 Changelog
 ---------
 
+v1.12.0 (??.0?.2020)
+ * added support for TYPO3 10.4 LTS
+ * dropped support for TYPO3 6.2 LTS
+ * #165 wip refactoring of classes to psr4  
+
 v1.11.4 (13.04.2020)
  * remove 6.2 code from BE-ToolBox
  * fix clipboard link in ToolBox